string_nconcat

+++ 0x0C-more_malloc_free/1-string_nconcat.c
@@ -29,7 +29,7 @@ char *string_nconcat(char *s1, char *s2, unsigned int n)
 
 	if (output == NULL)
 		return (NULL);
-	if (n >= s2len) 
+	if (n >= (unsigned int)s2len)
 	{
 		for (i = 0; s1[i] != '\0'; i++)
 			output[i] = s1[i];
